replace existing treeview control
