--- packages/dw-react-controls/src/components/AccessibleTreeView/AccessibleTreeView.test.tsx
@@ -1,13 +1,0 @@
-import { render } from "@testing-library/react";
-import * as React from "react";
-import { AccessibleTreeView } from "./AccessibleTreeView";
-
-describe("<AccessibleTreeView />", () => {
-	describe("Rendering", () => {
-		it("should render empty control without crashing", () => {
-			const { getByRole } = render(<AccessibleTreeView />);
-
-			expect(getByRole(/tree/i)).toBeTruthy();
-		});
-	});
-});
