--- packages/dw-react-controls/src/components/AccessibleTreeView/AccessibleTreeView.tsx
@@ -1,15 +1,0 @@
-import { styled } from "office-ui-fabric-react/lib/Utilities";
-import * as React from "react";
-import { AccessibleTreeViewBase } from "./AccessibleTreeView.base";
-import { getStyles } from "./AccessibleTreeView.styles";
-import {
-	AccessibleTreeViewProps,
-	IAccessibleTreeViewStyleProps,
-	IAccessibleTreeViewStyles
-} from "./AccessibleTreeView.types";
-
-export const AccessibleTreeView: React.FC<AccessibleTreeViewProps> = styled<
-	AccessibleTreeViewProps,
-	IAccessibleTreeViewStyleProps,
-	IAccessibleTreeViewStyles
->(AccessibleTreeViewBase, getStyles);
