--- packages/dw-react-controls/src/components/AccessibleTreeView/index.ts
@@ -1,5 +1,0 @@
-export * from "./AccessibleTreeView";
-export * from "./AccessibleTreeView.base";
-export * from "./AccessibleTreeView.context";
-export { default as AccessibleTreeViewContext } from "./AccessibleTreeView.context";
-export * from "./AccessibleTreeView.types";
