--- packages/dw-react-controls/src/components/AccessibleTreeView/stories/AccessibleTreeView.stories.mdx
@@ -1,21 +1,0 @@
-import { ArgsTable, Meta, Story, Canvas } from "@storybook/addon-docs";
-import * as stories from "./AccessibleTreeView.stories";
-import { AccessibleTreeView } from "../AccessibleTreeView";
-import { IAccessibleTreeViewContext } from "../AccessibleTreeView.context";
-import { AccessibleTreeItem } from "../../AccessibleTreeItem";
-
-<Meta title="Components/AccessibleTreeView" component={AccessibleTreeView} subcomponents={{ AccessibleTreeItem }} />
-
-# AccessibleTreeView
-
-An `AccessibleTreeView` represents a hierarchical list which can be navigated using the keyboard only.
-
-### Basic usage
-
-<Canvas withToolbar={true}>
-	<Story story={stories.Basic} name="Basic usage" />
-</Canvas>
-
-## Properties
-
-<ArgsTable components={{ AccessibleTreeView, AccessibleTreeItem }} />
