+++ packages/dw-react-controls/src/components/TreeItem/TreeItem.base.tsx
@@ -1,144 +1,145 @@
 /**
  * Control very much based on Material UI TreeItem control.
- * Source code at https://github.com/mui-org/material-ui/blob/master/packages/material-ui-lab/src/TreeItem/TreeItem.js
+ * Source code at https://github.com/mui/material-ui/blob/master/packages/mui-lab/src/TreeItem/TreeItem.js
  */
 
-import { IconButton } from "office-ui-fabric-react/lib/Button";
-import { ContextualMenu, IContextualMenuItem } from "office-ui-fabric-react/lib/ContextualMenu";
-import { Icon } from "office-ui-fabric-react/lib/Icon";
+import { useForkRef } from "@dlw-digitalworkplace/dw-react-utils";
 import { classNamesFunction, IRenderFunction } from "office-ui-fabric-react/lib/Utilities";
 import * as React from "react";
 import { composeRenderFunction } from "../../utilities";
-import TreeViewContext from "../TreeView/TreeView.context";
-import { TreeItemContent } from "./sections/TreeItemContent";
-import { ITreeItemContentProps } from "./sections/TreeItemContent.types";
+import { TreeViewContext } from "../TreeView";
+import { DescendantProvider, IDescendant, useDescendant } from "../TreeView/descendants";
 import { ITreeItemProps, ITreeItemStyleProps, ITreeItemStyles } from "./TreeItem.types";
+import { TreeItemContent } from "./TreeItemContent";
 
 const getClassNames = classNamesFunction<ITreeItemStyleProps, ITreeItemStyles>();
 
+const ownerDocument = (element: Node | null | undefined): Document => (element && element.ownerDocument) || document;
+
 export const TreeItemBase: React.FC<ITreeItemProps> = React.forwardRef<HTMLLIElement, ITreeItemProps>((props, ref) => {
 	const {
-		children,
-		nodeId,
-		label,
-		disabled,
-		iconName,
 		actions,
+		children,
+		disabled: disabledProp,
+		iconName,
+		id: idProp,
+		label,
+		nodeId,
 		onClick,
 		onInvoke,
-		onRenderItemContents,
-		styles,
-		className,
-		theme
+		onRenderContent,
+		onRenderLabel
 	} = props;
-	const { isExpanded, isSelected, selectNode, toggleExpansion } = React.useContext(TreeViewContext);
+	const { className, styles, theme } = props;
 
-	const contextMenuIconRef = React.useRef<HTMLDivElement>(null);
-	const [showContextMenu, setShowContextMenu] = React.useState(false);
+	const {
+		focus,
+		disabledItemsFocusable,
+		isDisabled,
+		isExpanded,
+		isFocused,
+		// isSelected,
+		mapFirstChar,
+		registerNode,
+		treeId,
+		unmapFirstChar,
+		unregisterNode
+	} = React.useContext(TreeViewContext);
 
-	const isExpandable = Array.isArray(children) ? !!children.length : !!children;
-	const expanded = isExpanded(nodeId);
-	const selected = isSelected(nodeId);
+	let id = null;
 
-	const handleExpandIconClick = (event: React.MouseEvent<HTMLElement>): void => {
-		event.stopPropagation();
+	if (!!idProp) {
+		id = idProp;
+	} else if (treeId && nodeId) {
+		id = `${treeId}-${nodeId}`;
+	}
 
-		if (isExpandable) {
-			toggleExpansion(event, nodeId);
+	const classNames = getClassNames(styles, { className, theme: theme! });
+
+	const [treeItemElement, setTreeItemElement] = React.useState<HTMLLIElement | null>(null);
+	const handleRef = useForkRef(setTreeItemElement, ref);
+
+	const descendant = React.useMemo<IDescendant>(
+		() => ({
+			element: treeItemElement,
+			id: nodeId
+		}),
+		[nodeId, treeItemElement]
+	);
+	const { index, parentId } = useDescendant(descendant);
+
+	const disabled = isDisabled ? isDisabled(nodeId) : false;
+	const expandable = Array.isArray(children) ? children.length > 0 : !!children;
+	const expanded = isExpanded ? isExpanded(nodeId) : false;
+	const focused = isFocused ? isFocused(nodeId) : false;
+	// const selected = isSelected ? isSelected(nodeId) : false;
+
+	React.useEffect(() => {
+		if (index !== -1) {
+			registerNode({
+				id: nodeId,
+				index,
+				parentId,
+				expandable,
+				disabled: !!disabledProp
+			});
+
+			return () => unregisterNode(nodeId);
+		}
+
+		return undefined;
+	}, [registerNode, unregisterNode, parentId, index, nodeId, expandable]);
+
+	React.useEffect(() => {
+		if (!!label) {
+			mapFirstChar(nodeId, label.substring(0, 1));
+
+			return () => {
+				unmapFirstChar(nodeId);
+			};
+		}
+		return undefined;
+	}, [mapFirstChar, unmapFirstChar, nodeId, label]);
+
+	const handleFocus = (event: React.FocusEvent<HTMLElement>) => {
+		if (event.target === event.currentTarget) {
+			ownerDocument(event.target).getElementById(treeId)?.focus({ preventScroll: true });
+		}
+
+		const unfocusable = !disabledItemsFocusable && disabled;
+		if (!focused && event.target === event.currentTarget && !unfocusable) {
+			focus(event, nodeId);
 		}
 	};
 
-	const handleExpandIconDoubleClick = (event: React.MouseEvent<HTMLElement>): void => {
-		event.stopPropagation();
+	const defaultRenderContent: IRenderFunction<ITreeItemProps> = (contentProps: ITreeItemProps) => {
+		return (
+			<TreeItemContent
+				actions={actions}
+				iconName={iconName}
+				label={label}
+				nodeId={nodeId}
+				onClick={onClick}
+				onInvoke={onInvoke}
+				onRenderLabel={onRenderLabel}
+			/>
+		);
 	};
 
-	const handleClick = (event: React.MouseEvent<HTMLElement>): void => {
-		if (!disabled) {
-			selectNode(event, nodeId);
-		}
-
-		if (onClick) {
-			onClick(event);
-		}
-	};
-
-	const handleDoubleClick = (event: React.MouseEvent<HTMLElement>): void => {
-		if (onInvoke) {
-			onInvoke(event);
-		}
-	};
-
-	const handleContextMenuButtonClick = (_: React.MouseEvent<HTMLDivElement>): void => {
-		setShowContextMenu(true);
-	};
-
-	const handleHideContextMenu = (_?: React.MouseEvent | React.KeyboardEvent): void => {
-		setShowContextMenu(false);
-	};
-
-	const contextMenuActions: IContextualMenuItem[] =
-		actions?.map((action) => ({
-			...action,
-			onClick: () => (typeof action.onClick === "function" ? action.onClick(nodeId) : false)
-		})) || [];
-
-	const classNames = getClassNames(styles, { className, disabled, expanded, selected, theme: theme! });
-
-	const renderItemContents: IRenderFunction<ITreeItemContentProps> = (
-		contentProps: ITreeItemContentProps | undefined
-	) => {
-		return contentProps ? <TreeItemContent {...contentProps} /> : null;
-	};
-
-	const finalOnRenderItemContents = onRenderItemContents
-		? composeRenderFunction(onRenderItemContents, renderItemContents)
-		: renderItemContents;
+	const renderContent = onRenderContent
+		? composeRenderFunction(onRenderContent, defaultRenderContent)
+		: defaultRenderContent;
 
 	return (
-		<li
-			ref={ref}
-			role={"treeitem"}
-			className={classNames.treeItem}
-			aria-expanded={isExpandable ? expanded : undefined}
-			aria-selected={selected}
-		>
-			<div className={classNames.treeNode} onClick={handleClick} onDoubleClick={handleDoubleClick}>
-				<div className={classNames.expandIconWrapper}>
-					{isExpandable && (
-						<Icon
-							className={classNames.expandIcon}
-							iconName={expanded ? "ChevronDownSmall" : "ChevronRightSmall"}
-							onClick={handleExpandIconClick}
-							onDoubleClick={handleExpandIconDoubleClick}
-						/>
-					)}
-				</div>
-
-				{finalOnRenderItemContents({ disabled, selected, expanded, label, iconName })}
-
-				{actions && actions.length > 0 && (
-					<div ref={contextMenuIconRef} className={classNames.contextMenuIconWrapper}>
-						<IconButton
-							className={classNames.contextMenuIcon}
-							onClick={handleContextMenuButtonClick}
-							iconProps={{ iconName: "More" }}
-						/>
-					</div>
-				)}
-
-				{showContextMenu && (
-					<ContextualMenu
-						items={contextMenuActions}
-						onDismiss={handleHideContextMenu}
-						target={contextMenuIconRef.current}
-					/>
-				)}
-			</div>
+		<li className={classNames.root} id={id!} onFocus={handleFocus} ref={handleRef} role={"TreeItem"}>
+			{renderContent(props)}
 
 			{expanded && children && (
-				<ul role={"group"} className={classNames.childNodes}>
-					{children}
-				</ul>
+				<DescendantProvider id={nodeId}>
+					<ul role="group" className={classNames.children}>
+						{children}
+					</ul>
+				</DescendantProvider>
 			)}
 		</li>
 	);
