+++ packages/dw-react-controls/src/components/TreeItem/TreeItem.styles.ts
@@ -1,30 +1,18 @@
-import { AnimationClassNames } from "office-ui-fabric-react/lib/Styling";
 import { ITreeItemStyleProps, ITreeItemStyles } from "./TreeItem.types";
+import { expandIconWidth } from "./TreeItemContent/TreeItemContent.styles";
 
 const GlobalClassNames = {
-	treeItem: "dw-TreeItem",
-	treeNode: "dw-TreeNode",
-	contextMenuIconWrapper: "dw-TreeItemContextMenuIconWrapper",
-	contextMenuIcon: "dw-TreeItemContextMenuIcon",
-	expandIconWrapper: "dw-TreeExpandIconWrapper",
-	expandIcon: "dw-TreeExpandIcon",
-	childNodes: "dw-TreeNodeChildren",
-	itemWrapper: "dw-TreeNodeItemWrapper",
-	labelWrapper: "dw-TreeNodeLabelWrapper",
-	label: "dw-TreeNodeLabel",
-	iconWrapper: "dw-TreeNodeIconWrapper",
-	icon: "dw-TreeNodeIcon"
+	root: "dw-TreeItem",
+	children: "dw-TreeItem-children"
 };
 
-const expandIconWidth = "22px";
-
 export const getStyles = (props: ITreeItemStyleProps): ITreeItemStyles => {
-	const { className, disabled, expanded, selected } = props;
+	const { className } = props;
 	const classNames = GlobalClassNames;
 
 	return {
-		treeItem: [
-			classNames.treeItem,
+		root: [
+			classNames.root,
 			{
 				boxShadow: "border-box",
 				listStyleType: "none",
@@ -32,53 +20,14 @@ export const getStyles = (props: ITreeItemStyleProps): ITreeItemStyles => {
 				padding: 0,
 				userSelect: "none"
 			},
-			disabled && "is-disabled",
-			expanded && "is-expanded",
-			selected && "is-selected",
 			className
 		],
 
-		treeNode: [
-			classNames.treeNode,
+		children: [
+			classNames.children,
 			{
-				alignItems: "center",
-				display: "flex",
-				height: "30px",
-				selectors: {
-					":hover": {
-						cursor: disabled ? "not-allowed" : "pointer"
-					}
-				}
+				paddingInlineStart: expandIconWidth
 			}
-		],
-
-		contextMenuIconWrapper: [
-			classNames.contextMenuIconWrapper,
-			{
-				marginLeft: selected ? "3px" : "4px"
-			}
-		],
-
-		contextMenuIcon: [classNames.contextMenuIcon],
-
-		expandIconWrapper: [
-			classNames.expandIconWrapper,
-			{
-				width: expandIconWidth
-			}
-		],
-		expandIcon: [
-			classNames.expandIcon,
-			{
-				padding: "4px",
-				selectors: {
-					":hover": {
-						cursor: "pointer"
-					}
-				}
-			}
-		],
-
-		childNodes: [classNames.childNodes, { paddingLeft: expandIconWidth }, AnimationClassNames.fadeIn100]
+		]
 	};
 };
