+++ packages/dw-react-controls/src/components/TreeItem/TreeItem.types.ts
@@ -1,32 +1,47 @@
 import { IStyle, ITheme } from "office-ui-fabric-react/lib/Styling";
 import { IRenderFunction, IStyleFunctionOrObject } from "office-ui-fabric-react/lib/Utilities";
-import { ITreeItemContentProps } from "./sections/TreeItemContent.types";
+import { ITreeItemAction } from "../TreeItem";
+import { ITreeItemContentProps } from "./TreeItemContent";
 
 export interface ITreeItemProps {
-	/**
-	 * The unique id of this tree item
-	 */
-	nodeId: string;
-
-	/**
-	 * The label to show for the tree item
-	 */
-	label: string;
-
 	/**
 	 * Actions which are available on the tree item
 	 */
 	actions?: ITreeItemAction[];
 
-	/**
-	 * Controls whether the tree item can be selected
-	 */
 	disabled?: boolean;
 
-	/**
-	 * Specifies the icon name to show with the term label
-	 */
 	iconName?: string;
+
+	id?: string;
+
+	label?: string;
+
+	nodeId: string;
+
+	/**
+	 * Callback fired when a the content is clicked
+	 *
+	 * @param {object} event The event source of the callback.
+	 */
+	onClick?: (event: React.SyntheticEvent) => void;
+
+	/**
+	 * Callback fired when a the node is invoked
+	 *
+	 * @param {object} event The event source of the callback.
+	 */
+	onInvoke?: (event: React.SyntheticEvent) => void;
+
+	/**
+	 * When specified it will override the default rendering of the entire node
+	 */
+	onRenderContent?: IRenderFunction<ITreeItemProps>;
+
+	/**
+	 * When specified it will override the default rendering of the node label
+	 */
+	onRenderLabel?: IRenderFunction<ITreeItemContentProps>;
 
 	/**
 	 * Optional class for the root TreeItem element
@@ -42,48 +57,14 @@ export interface ITreeItemProps {
 	 * The current theme applied to the control
 	 */
 	theme?: ITheme;
-
-	/**
-	 * When specified it will override the default rendering of the TreeItemContent section
-	 */
-	onRenderItemContents?: IRenderFunction<ITreeItemContentProps>;
-
-	/**
-	 * Optional callback when the tree node has been clicked
-	 */
-	onClick?(event: React.MouseEvent<HTMLElement>): void;
-
-	/**
-	 * Optional callback when the tree node has been invoked
-	 */
-	onInvoke?(event: React.MouseEvent<HTMLElement>): void;
 }
 
 export interface ITreeItemStyleProps {
 	className?: string;
-	disabled?: boolean;
-	expanded?: boolean;
-	selected?: boolean;
 	theme: ITheme;
 }
 
 export interface ITreeItemStyles {
-	treeItem?: IStyle;
-	treeNode?: IStyle;
-	contextMenuIconWrapper?: IStyle;
-	contextMenuIcon?: IStyle;
-	expandIconWrapper?: IStyle;
-	expandIcon?: IStyle;
-	itemWrapper?: IStyle;
-	iconWrapper?: IStyle;
-	icon?: IStyle;
-	labelWrapper?: IStyle;
-	label?: IStyle;
-	childNodes?: IStyle;
-}
-
-export interface ITreeItemAction {
-	key: string;
-	text: string;
-	onClick?(nodeId: string): void;
+	root?: IStyle;
+	children?: IStyle;
 }
