+++ packages/dw-react-controls/src/components/TreeView/TreeView.context.ts
@@ -1,10 +1,22 @@
 import * as React from "react";
+import { ITreeViewNode } from "./TreeView.types";
 
 export interface ITreeViewContext {
-	isExpanded(id: string): boolean;
-	isSelected(id: string): boolean;
-	selectNode(event: React.SyntheticEvent<HTMLElement>, id: string): void;
-	toggleExpansion(event: React.SyntheticEvent<HTMLElement>, id: string): void;
+	disabledItemsFocusable?: boolean;
+	focus: (event: React.SyntheticEvent, nodeId: string) => void;
+	isDisabled: (nodeId: string) => boolean;
+	isExpandable: (nodeId: string) => boolean;
+	isExpanded: (nodeId: string) => boolean;
+	isFocused: (nodeId: string) => boolean;
+	isSelected: (nodeId: string) => boolean;
+	mapFirstChar: (nodeId: string, firstChar: string) => void;
+	multiSelect?: boolean;
+	registerNode: (node: ITreeViewNode) => void;
+	selectNode: (event: React.SyntheticEvent, nodeId: string | null, multiple?: boolean) => void;
+	toggleExpansion: (event: React.SyntheticEvent, nodeId?: string) => void;
+	treeId: string;
+	unmapFirstChar: (nodeId: string) => void;
+	unregisterNode: (nodeId: string) => void;
 }
 
 const TreeViewContext = React.createContext<ITreeViewContext>({} as ITreeViewContext);
