+++ packages/dw-react-controls/src/components/TreeView/TreeView.tsx
@@ -2,9 +2,9 @@ import { styled } from "office-ui-fabric-react/lib/Utilities";
 import * as React from "react";
 import { TreeViewBase } from "./TreeView.base";
 import { getStyles } from "./TreeView.styles";
-import { ITreeViewProps, ITreeViewStyleProps, ITreeViewStyles } from "./TreeView.types";
+import { ITreeViewStyleProps, ITreeViewStyles, TreeViewProps } from "./TreeView.types";
 
-export const TreeView: React.FC<ITreeViewProps> = styled<ITreeViewProps, ITreeViewStyleProps, ITreeViewStyles>(
+export const TreeView: React.FC<TreeViewProps> = styled<TreeViewProps, ITreeViewStyleProps, ITreeViewStyles>(
 	TreeViewBase,
 	getStyles
 );
