+++ packages/dw-react-controls/src/components/TreeView/stories/TreeView.stories.mdx
@@ -8,27 +8,12 @@ import { TreeItem } from "../../TreeItem";
 
 # TreeView
 
-A `TreeView` represents a hierarchical list.
+An `TreeView` represents a hierarchical list which can be navigated using the keyboard only.
 
 ### Basic usage
 
 <Canvas withToolbar={true}>
 	<Story story={stories.Basic} name="Basic usage" />
-</Canvas>
-
-### Advanced usage (uncontrolled)
-
-<Canvas withToolbar={true}>
-	<Story story={stories.AdvancedUncontrolled} name="Advanced usage (uncontrolled)" />
-</Canvas>
-
-### Advanced usage (controlled)
-
-Using the **controlled** version of the component allows you to control the `expanded` and `selected` items
-of the `TreeView`.
-
-<Canvas withToolbar={true}>
-	<Story story={stories.AdvancedControlled} name="Advanced usage (controlled)" />
 </Canvas>
 
 ## Properties
